host/volume: Create volume.Manager outline.
This class will be responsible for central state keeping about volumes in an active host daemon; though most of the work will be delegated out to backend implementations, this will be responsible for enumerating those and making sure there's a readily accessible list of all volumes.

Start massaging structures to clearly mark which ones are serialable for both upcoming API and persistence usage.

Signed-off-by: Eric Myhre <eric@flynn.io>