router: add a cheap way to handle sticky sessions
In the new net/http path, use a special to mark whether the service
should use sticky sessions on a request. In the service, check for &
delete this header, and use its presence to determine whether to use the
sticky session picker.

Make sure to set the sticky session cookie after the response is
received from the backend.

Signed-off-by: Blake Gentry <blakesgentry@gmail.com>